requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
cfgrib==0.9.11.0
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
eccodes==1.7.0
findlibs==0.0.5
fonttools==4.51.0
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.8.4
netCDF4==1.6.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pygrib==2.1.5
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
xarray==2024.5.0
